Create separate exception for json5 load error
